summary
I have added resore option to the file. you can use it by changing RESTORE  to 1, if you do so it will restore the run that found at /variables/unet_3_ddmmyy
if you want to restore from specific date than fill the varibale DATE with the desired date in the format of ddmmyy else it will restore from todays date

note that i have deleted the ExternalModules.network_skeleton in before the package name in the import section because it gave me errors.